refactor(language-service): Move some util functions to common
These functions will be useful to the Ivy language service as well to
provide 'go to definition' functionality for template and style urls.
