+++ packages/language-service/common/BUILD.bazel
@@ -6,6 +6,7 @@ ts_library(
     name = "common",
     srcs = glob(["*.ts"]),
     deps = [
+        "@npm//@types/node",
         "@npm//typescript",
     ],
 )
